datasource: more robust file naming
To avoid collisions when multiple datasources have files with the same
name, the constructed RDFox working directory now uses a hash of all the
files contained in the data source to create the subdirectory.

This folder can be accessed using the `dir.datasource` RDFox shell
variable when loading data sources.